communication.hpp

+++ algorithms/src/communication.hpp
@@ -1,7 +1,7 @@
 
 struct locate_motor{
-double motor_right;
-double motor_left;
+int motor_right;
+int motor_left;
 };
 
 locate_motor get_desired_motor();
